requirements.txt
requests
beautifulsoup4
google-play-scraper
openai
pillow
pure-python-adb
xmltodict
numpy
opencv-python
torch
torchvision
tiktoken